App now checks user id in edit/remove

+++ app.py
@@ -1,6 +1,6 @@
 import sqlite3
 from flask import Flask
-from flask import redirect, render_template, request, session
+from flask import abort, redirect, render_template, request, session
 from werkzeug.security import check_password_hash, generate_password_hash
 import config
 import db
@@ -34,11 +34,16 @@ def create_item():
 @app.route("/edit_item/<int:item_id>")
 def edit_item(item_id):
     item = items.get_item(item_id)
+    if item["user_id"] != session["user_id"]:
+        abort(403)
     return render_template("edit_item.html", item=item)
 
 @app.route("/update_item", methods=["POST"])
 def update_item():
     item_id = request.form["item_id"]
+    item =items.get_item(item_id)
+    if item["user_id"] != session["user_id"]:
+        abort(403)
     title = request.form["title"]
     description = request.form["description"]
     items.update_item(item_id, title, description)
@@ -46,8 +51,10 @@ def update_item():
 
 @app.route("/delete_item/<int:item_id>", methods=["GET", "POST"])
 def delete_item(item_id):
+    item = items.get_item(item_id)
+    if item["user_id"] != session["user_id"]:
+        abort(403)
     if request.method == "GET":
-        item = items.get_item(item_id)
         return render_template("delete_item.html", item=item)
     
     if request.method == "POST":
